* Arkbot  - Support for cross-wiki requests  - New module replica to copy a page across different wikis (eg. for iKiwi updates)  - * Minibot  - New module replica, won't be used much since Arkbot already does the job * Compression  - Added Rice encoding * Utilities  - Downloading all available dumps

+++ utilities/cleanDumps.sh
@@ -1,10 +1,14 @@
 #! /bin/sh
 
-for lang in aa ab af am ar as ay ba be bg bi bn bo br ca co cs cy da de el en es es et eu fa fi fj fo fr fy ga gd gl gn ha he hi hr hu hy id is it iu ja jv ka kk kl km kn ko ks ku la lb ln lo lt lv mg mi mk ml mn mo mr ms mt my ne nl no oc om or pa pl ps pt qu rn ro ru rw sa sd sg sh si sk sl sm sn so sq ss su sv sw ta te tg th ti tk tl to tr ts tt tw ug uk ur uz vi wo xh yi yo za zh zu; do
-    echo cleaning $lang
-    ~/local/bin/extract-wiki -b 1T < ${lang}wiki.xml
-    mv AA/wiki00 ${lang}wiki.txt
+for lang in aa ab ace af ak als am an ang ar arc arz as ast av ay az ba bar bat-smg bcl be be-x-old bg bh bi bm bn bo bpy br bs bug bxr ca cbk-zam cdo ce ceb ch chr chy ckb co cr crh cs csb cu cv cy da de diq dsb dv dz ee el eml en eo es et eu ext fa ff fi fiu-vro fj fo fr frp fur fy ga gan gd gl glk gn got gu gv ha hak haw he hi hif hr hsb ht hu hy ia id ie ig ik ilo io is it iu ja jbo jv ka kaa kab kg ki kk kl km kn ko ks ksh ku kv kw ky la lad lb lbe lg li lij lmo ln lo lt lv map-bms mdf mg mhr mi mk ml mn mr ms mt mwl my myv mzn na nah nap nds nds-nl ne new ng nl nn no nov nrm nv ny oc om or os pa pag pam pap pcd pdc pi pih pl pms pnb pnt ps pt qu rm rmy rn ro roa-rup roa-tara ru rw sa sah sc scn sco sd se sg sh si simple sk sl sm sn so sq sr srn ss st stq su sv sw szl ta te tet tg th ti tk tl tn to tpi tr ts tt tum tw ty udm ug uk ur uz ve vec vi vls vo wa war wo wuu xal xh yi yo za zea zh zh-classical zh-min-nan zh-yue zu; do
+    if ! [ -f txt/${lang}wiki.txt ]; then
+	echo cleaning $lang
+        ~/local/bin/extract-wiki -b 1T < xml/${lang}wiki.xml
+        mv AA/wiki00 txt/${lang}wiki.txt
+    else
+	echo already cleaned $lang
+    fi
 done
-rmdir AA
+[ -d AA ] && rmdir AA
 
 echo finished
